Do the components and also do the files organization, but need to improve that
