+++ src/core/components/Footer.jsx
@@ -1,0 +1,7 @@
+export const Footer = ()=>{
+    return(
+        <footer className='p-2 bg-azulFondo'>
+
+        </footer>
+    )
+}
